Add detailed logging to debug form submission issues
- Added console logging to both microplastic and PFA POST endpoints
- Logs request body, headers, and any errors for debugging
- Fixed salt field validation to accept decimals (matching DB schema)
- Added error details to API responses

This will help identify why form submissions are failing in production.

+++ client/src/components/dashboard/weekly-input-form.tsx
@@ -19,7 +19,7 @@ interface WeeklyInputFormProps {}
 const formSchema = z.object({
   bottledWater: z.number().int().min(0).default(0),
   seafood: z.number().int().min(0).default(0),
-  salt: z.number().int().min(0).default(0),
+  salt: z.number().min(0).default(0),
   plasticPackaged: z.number().int().min(0).default(0),
   teaBags: z.number().int().min(0).default(0),
   householdDust: z.number().int().min(0).default(0),
